+++ src/main/java/uk/gov/ch/corporatebody/service/CorporateBodyService.java
@@ -4,5 +4,5 @@ import uk.gov.ch.corporatebody.exception.CorporateBodyNotFoundException;
 
 public interface CorporateBodyService {
 
-    int getActionCode(String companyNumber) throws CorporateBodyNotFoundException;
+    long getActionCode(String companyNumber) throws CorporateBodyNotFoundException;
 }
